Audio rebalancing to more closely approximate the original Famicom HVC-001 model.
Huge thanks to the user riggles1 for this!

Reviewed the audio mixing in comparison with real hardware audio capture (HVC-001), Mesen currently sounds more similar to the HVC-101, better known as the AV Famicom. Meaning that the VRC6 volume is greatly overpowering the other channels (many times harmonies or additional background melody being inaudible) as well as a very weak DMC Sample channel making the drums hardly present. Both of these volume levels were tweaked in this pull request.
Soundtracks like Akumajou Densetsu were composed with the original model 001 mixing in mind.

I adjusted the mixing to arrive at the following results showcased in this video. With the changes (first) and what the Mesen core currently sounds like (second).
https://www.youtube.com/watch?v=gxgKe__TCgo
The VRC6 sawtooth bass used to be so loud that it was also popping a bit.

Real hardware capture examples can be heard in the RetroRGB article regarding this:
https://www.retrorgb.com/av-famicom-hvc-101-audio-balance-restoration-mod.html

This doesn't impact other games negatively, it especially helps music that utilizes a lot of drum or bass samples, techniques heavily used by developers at Konami and Sunsoft for example.